Moved ScriptTextEdit into separate module and reuse it in renaming
